requirements.txt
streamlit
pandas
scikit-learn
joblib
matplotlib
seaborn
gitpython
pytest
ruff
